+++ components/collector/tests/source_collectors/file_source_collectors/test_performancetest_runner.py
@@ -46,8 +46,8 @@ class PerformanceTestRunnerSlowTransactionsTest(PerformanceTestRunnerTestCase):
     async def test_warning_only(self):
         """Test that only transactions that exceed the warning threshold are counted."""
         html = '<html><table class="details"><tr class="transaction"><td class="red evaluated"/>' \
-            '</tr><tr class="transaction"><td class="name">Name</td><td class="yellow evaluated"/></tr>' \
-            '<tr class="transaction"><td class="green evaluated"/></tr></table></html>'
+               '</tr><tr class="transaction"><td class="name">Name</td><td class="yellow evaluated"/></tr>' \
+               '<tr class="transaction"><td class="green evaluated"/></tr></table></html>'
         self.sources["source_id"]["parameters"]["thresholds"] = ["warning"]
         response = await self.collect(self.metric, get_request_text=html)
         self.assert_measurement(response, value="1", entities=[dict(key="Name", name="Name", threshold="warning")])
@@ -60,6 +60,17 @@ class PerformanceTestRunnerSlowTransactionsTest(PerformanceTestRunnerTestCase):
                '<tr class="transaction"><td class="name">T3</td><td class="green evaluated"/></tr>' \
                '</table></html>'
         self.sources["source_id"]["parameters"]["transactions_to_ignore"] = ["T[1|3]"]
+        response = await self.collect(self.metric, get_request_text=html)
+        self.assert_measurement(response, value="1", entities=[dict(key="T2", name="T2", threshold="warning")])
+
+    async def test_include_transactions_by_name(self):
+        """Test that transactions can be included by name."""
+        html = '<html><table class="details">' \
+               '<tr class="transaction"><td class="name">T1</td><td class="red evaluated"/></tr>' \
+               '<tr class="transaction"><td class="name">T2</td><td class="yellow evaluated"/></tr>' \
+               '<tr class="transaction"><td class="name">T3</td><td class="green evaluated"/></tr>' \
+               '</table></html>'
+        self.sources["source_id"]["parameters"]["transactions_to_include"] = ["T2"]
         response = await self.collect(self.metric, get_request_text=html)
         self.assert_measurement(response, value="1", entities=[dict(key="T2", name="T2", threshold="warning")])
 
@@ -103,13 +114,13 @@ class PerformanceTestRunnerTestsTest(PerformanceTestRunnerTestCase):
 </html>'''
 
     async def test_tests(self):
-        """Test that the number of performancetest transactions is returned."""
+        """Test that the number of performance test transactions is returned."""
         metric = dict(type="tests", sources=self.sources, addition="sum")
         response = await self.collect(metric, get_request_text=self.html)
         self.assert_measurement(response, value="28", total="28")
 
     async def test_failed_tests(self):
-        """Test that the number of failed performancetest transactions is returned."""
+        """Test that the number of failed performance test transactions is returned."""
         # We also pass an obsolete status ("canceled") to test that obsolete statuses are ignored:
         self.sources["source_id"]["parameters"]["test_result"] = ["failed", "canceled"]
         metric = dict(type="tests", sources=self.sources, addition="sum")
@@ -117,27 +128,34 @@ class PerformanceTestRunnerTestsTest(PerformanceTestRunnerTestCase):
         self.assert_measurement(response, value="8", total="28")
 
     async def test_succeeded_tests(self):
-        """Test that the number of succeeded performancetest transactions is returned."""
+        """Test that the number of succeeded performance test transactions is returned."""
         self.sources["source_id"]["parameters"]["test_result"] = ["success"]
         metric = dict(type="tests", sources=self.sources, addition="sum")
         response = await self.collect(metric, get_request_text=self.html)
         self.assert_measurement(response, value="20", total="28")
 
     async def test_ignored_tests(self):
-        """Test that the number of performancetest transactions is returned for transactions that are not ignored."""
+        """Test that the number of performance test transactions is returned for transactions that are not ignored."""
         self.sources["source_id"]["parameters"]["transactions_to_ignore"] = [".*2"]
         metric = dict(type="tests", sources=self.sources, addition="sum")
         response = await self.collect(metric, get_request_text=self.html)
         self.assert_measurement(response, value="10", total="10")
 
     async def test_failed_and_ignored_tests(self):
-        """Test that the number of failed performancetest transactions is returned for transactions that are not
+        """Test that the number of failed performance test transactions is returned for transactions that are not
         ignored."""
         self.sources["source_id"]["parameters"]["transactions_to_ignore"] = [".*2"]
         self.sources["source_id"]["parameters"]["test_result"] = ["failed"]
         metric = dict(type="tests", sources=self.sources, addition="sum")
         response = await self.collect(metric, get_request_text=self.html)
         self.assert_measurement(response, value="3", total="10")
+
+    async def test_included_tests(self):
+        """Test that the number of performance test transactions is returned for transactions that are included."""
+        self.sources["source_id"]["parameters"]["transactions_to_include"] = ["T1"]
+        metric = dict(type="tests", sources=self.sources, addition="sum")
+        response = await self.collect(metric, get_request_text=self.html)
+        self.assert_measurement(response, value="10", total="10")
 
 
 class PerformanceTestRunnerStabilityTest(PerformanceTestRunnerTestCase):